Even more debugging
I've caught a CAS failure live (when attempting to fail a job).
Contrary to my expectations, it wasn't lock_version that didn't match,
it was the status that was expected to be :started (1) but was
actually :unstarted (0).

So the question is, how did it come to be :unstarted? Either
reserve-job was never called (or its transaction never completed), or
something at some point had called reset-job (it shouldn't). It's
unclear from the logs I have, so here's another commit to gather
more forensic evidence.