And even more debugging
The plot thickens. In another run against the previous commit, I've
been able to track down the CAS failure to the monitor detecting a job
with failed heartbeats, assuming it's dead, and proceeding to reset
its status to :unstarted. This would explain the status mismatch.

What happened in this case is

- the heartbeat thread continuously emitted heartbeat log messages
  every minute
- at some point they stopped
- 8 mins later, monitor noticed and said "Found 1 jobs with failed
  heartbeats"

So, then, the question is: why did heartbeats stop? I see two possible
explanations:

- either something else than Exception was thrown in
  `core/heartbeat-job`, causing the future to yield an ExecutionException
  when eventually derefed (but it never is); or
- `core/heartbeat-job` hangs indefinitely (SQLite lockup?)

To detect the first case, I've replaced `catch Exception` with `catch
Throwable` so that we catch everything. To detect the second, I've
added more logging around executing the SQL in `heartbeat-job`.