More more more
In the previous commit I formulated two hypotheses, and none of them
seems to be true; the heartbeat tx seems to complete, but then it
doesn't resume on the next iteration.

Let's resort to dumping stacktraces: see what the heartbeat thread
is doing (if anything) when the monitor detects failed heartbeats.